client/requirements.txt
click
google-auth
grpcio
protobuf
pyarrow
